Ignore "--" during usage block parsing
This is parsing as an option (not sure if `:long-option` or
`:short-option`, however it appears in result and it is never true).

So create its own token and ignore it.

As you can see in the "Complex command" test case, this is still wrong
in one more complex case. But at least it accepts the grammar now
(before it outright rejected it).