Better fix the issue with `--`
This is a partial revert of commits:
- d2413b2ec95203fbc452ef4fbfe4d0294151dfc0 (ignore `--` in
  docopt.usageblock/tokenize-pattern)
- 6845163cbdd1ba8756905d5ac1b97de85811f603 (find biggest-match in
  docopt.match/match-argv)

The actual root cause of the issue while parsing `--` was the
`docopt.match/parse` function. `docopt.usageoption/parse` seems correct,
at least while parsing this:

```
(parse "usage: prog foo -- <extra-opts>...")

:docopt.usageblock/required
[ :docopt.usageblock/command "foo" ]
[ :docopt.usageblock/command "--" ]
[ :docopt.usageblock/repeat [ :docopt.usageblock/argument "<extra-opts>" ] ]
```

As it is clear above, `--` is correctly parsed as a
`:docopt.usageblock/command`. This is the same behavior as
docopt/docopt, but commit d2413b2ec95203fbc452ef4fbfe4d0294151dfc0 broke
this.

Reverting commit d2413b2ec95203fbc452ef4fbfe4d0294151dfc0 and now the
issue is in `docopt.match/parse`, since it was ignoring `--` thanks to
the `[head _ & tail] (partition-by (partial "--") argv)` part. Removing
`_` and putting `--` at the tail seems to do the trick.

Compared to the previous trick, this fix some tests and breaks other
ones. But I think the code as a whole now is more correctly, and the
other issues can be fixed by other commits.