Add docopt.match/best-match-by-argv
This is a hack similar to the previous biggest-match from commit
6845163cbdd1ba8756905d5ac1b97de85811f603, but now that `--` is parsed
correctly again by `docopt.usageblock/parse`, this is a simpler hack.

All it does is, if there is a `--` parsed in argv AND there is a
possible match with `--`, it probably means this is the correct match.

Fixes almost all remaining (known) issues with this, with exception of
parsing `[--]`.