Workaround issues with quoted arguments with spaces
The parse has an issue that when we have something like:

```clojure
(docopt "usage: prog <foo>" ["a b"])
```

It will try to parse "a" and "b" as 2 separate arguments (and this will
obviously fail).

Fixing the parse seems very hard, so let's do a workaround instead,
converting spaces to \u00A0 (non-breaking space). So for example,
conveting "a b" to "a\u00A0b".

Ugly, but it works.